+++ packages/components/lib/components/primitives/Tooltip/Tooltip.d.ts
@@ -1,5 +1,9 @@
 import React from 'react';
 export declare const TOOLTIP_DEFAULT_TEST_ID = "tooltip";
+export declare enum TooltipBackgroundColor {
+    Primary900 = "bg-fill-primary-900",
+    Primary100 = "bg-fill-primary-100"
+}
 export declare enum TooltipPlace {
     Left = "left",
     Right = "right",
@@ -7,14 +11,13 @@ export declare enum TooltipPlace {
     Top = "top"
 }
 declare type Props = {
-    id: string;
     place?: TooltipPlace;
-    tooltip: React.ReactNode;
     element: React.ReactNode;
-    testId?: string;
-    backgroundColor?: string;
+    tooltip: React.ReactNode;
     className?: string;
+    backgroundColor?: string;
+    testId?: string;
 };
-export declare const Tooltip: ({ id, place, tooltip, element, testId, backgroundColor, className, }: Props) => JSX.Element;
+export declare const Tooltip: ({ place, element, tooltip, className, backgroundColor, testId, }: Props) => JSX.Element;
 export {};
 //# sourceMappingURL=Tooltip.d.ts.map