+++ packages/components/lib/components/primitives/Tooltip/stories/Tooltip.stories.js
@@ -7,7 +7,7 @@ exports.base = exports.default = void 0;
 
 var _react = _interopRequireDefault(require("react"));
 
-var _ = require("..");
+var _Tooltip = require("../Tooltip");
 
 var _addonKnobs = require("@storybook/addon-knobs");
 
@@ -15,33 +15,17 @@ function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { de
 
 var _default = {
   title: 'Composer/Primitive/Tooltip',
-  component: _.Tooltip
+  component: _Tooltip.Tooltip
 };
 exports.default = _default;
-var idOptions;
-
-(function (idOptions) {
-  idOptions["top"] = "top";
-  idOptions["bottom"] = "bottom";
-  idOptions["left"] = "left";
-  idOptions["right"] = "right";
-})(idOptions || (idOptions = {}));
-
-var colorOptions;
-
-(function (colorOptions) {
-  colorOptions["red"] = "red";
-  colorOptions["gray"] = "gray";
-})(colorOptions || (colorOptions = {}));
 
 const base = () => /*#__PURE__*/_react.default.createElement("div", {
   className: "p-48"
-}, /*#__PURE__*/_react.default.createElement(_.Tooltip, {
-  id: (0, _addonKnobs.select)('id', idOptions, idOptions.top),
-  place: (0, _addonKnobs.select)('place', _.TooltipPlace, _.TooltipPlace.Top),
-  tooltip: /*#__PURE__*/_react.default.createElement("span", null, "Tooltip"),
-  element: /*#__PURE__*/_react.default.createElement("span", null, "Trigger"),
-  backgroundColor: (0, _addonKnobs.select)('backgroundColor', colorOptions, colorOptions.gray)
+}, /*#__PURE__*/_react.default.createElement(_Tooltip.Tooltip, {
+  place: (0, _addonKnobs.select)('Place', _Tooltip.TooltipPlace, _Tooltip.TooltipPlace.Left),
+  element: /*#__PURE__*/_react.default.createElement("span", null, (0, _addonKnobs.text)('Trigger', 'Trigger')),
+  tooltip: /*#__PURE__*/_react.default.createElement("span", null, (0, _addonKnobs.text)('Tooltip', 'Tooltip')),
+  backgroundColor: (0, _addonKnobs.select)('backgroundColor', _Tooltip.TooltipBackgroundColor, _Tooltip.TooltipBackgroundColor.Primary900)
 }));
 
 exports.base = base;