+++ packages/components/src/utils/getDropdownClasses/index.test.ts
@@ -11,7 +11,7 @@ describe('getDropdownClasses', () => {
     const itemClasses = getItemClasses();
     expect(itemClasses).toStrictEqual({
       wrapper:
-        'flex items-center justify-left px-6 py-4 hover:bg-state-ui-hover whitespace-no-wrap cursor-pointer focus:outline-none',
+        'flex items-center justify-left px-6 py-4 hover:bg-fill-state-ui-hover whitespace-no-wrap cursor-pointer focus:outline-none',
       iconClasses: 'fill-current mr-3',
       titleWrapper: 'flex flex-col',
     });